Get it to build with modern Rust
Updated dependencies, fixed some errors, formatting, fixed clippy
errors, refactoring / renaming. Functionality-wise it should be the
same.

I needed the Rust PGS parsing for a different project I'm working on, so
I thought I could use this one as a starting point. To get it to run I
had to do all of the above so thought I'd PR it back.